+++ packages/sdk-codemod/codemods/v3/setup-branch-flag-rename/tests/equals-form/input.sh
@@ -1,2 +1,0 @@
-tailor setup branch --name=my-app --branch=main
-tailor setup branch --branch='release' --force