+++ packages/sdk-codemod/codemods/v3/setup-branch-flag-rename/tests/line-continuation/expected.sh
@@ -1,8 +1,0 @@
-tailor setup branch \
-  --name my-app-stg \
-  --target main \
-  --environment staging
-tailor \
-  setup \
-  branch \
-  --target release