+++ packages/sdk/src/types/auth.generated.ts
@@ -1403,7 +1403,7 @@ export type AuthConfigInput =
           /** Field metadata configuration */
           metadata: {
             /** Date value representation */
-            as?: "string" | "date" | undefined;
+            as?: "string" | "date" | "temporal" | undefined;
             /** Whether the field is required */
             required?: boolean | undefined;
             /** Whether the field is an array */
@@ -2875,7 +2875,7 @@ export type AuthConfig =
           /** Field metadata configuration */
           metadata: {
             /** Date value representation */
-            as?: "string" | "date" | undefined;
+            as?: "string" | "date" | "temporal" | undefined;
             /** Whether the field is required */
             required?: boolean | undefined;
             /** Whether the field is an array */